# backend/docs/swagger/swagger.yaml
basePath: /
...
host: localhost:8088
info:
  contact: {}
  description: Bilingual CMS backend API for Impress (印迹). Supports content management,
    articles, pages, themes, media, and more.
  title: Impress CMS API
  version: "1.0"
paths:
...
  /admin/pages:
    get:
      description: Returns pages with optional status, parentId, and themeId filters
      parameters:
      - description: Status filter (draft/published)
        in: query
        name: status
        type: string
      - description: Parent page ID filter
        in: query
        name: parentId
        type: integer
      - description: Theme ID filter
        in: query
        name: themeId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              items:
                items:
                  type: object
                type: array
            type: object
      security:
      - BearerAuth: []
      summary: List all pages (admin)
      tags:
      - Pages (Admin)
    post:
      consumes:
      - application/json
      description: Create a new page (draft by default)
      parameters:
      - description: Page data
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create page
      tags:
      - Pages (Admin)
  /admin/pages/{id}:
    delete:
      description: Soft-delete a page by ID
      parameters:
      - description: Page ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete page
      tags:
      - Pages (Admin)
    get:
      description: Returns a single page by its database ID
      parameters:
      - description: Page ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get page by ID (admin)
      tags:
      - Pages (Admin)
    put:
      consumes:
      - application/json
      description: Update an existing page by ID
      parameters:
      - description: Page ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated page data
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update page
      tags:
      - Pages (Admin)
  /admin/pages/{id}/publish:
    put:
      description: Set a page's status to published
      parameters:
      - description: Page ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Publish page
      tags:
      - Pages (Admin)
  /admin/pages/{id}/unpublish:
    put:
      description: Revert a page's status back to draft
      parameters:
      - description: Page ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Unpublish page
      tags:
      - Pages (Admin)
...
  /public/pages:
    get:
      description: Returns all published pages with optional locale filtering
      parameters:
      - description: Locale (zh or en)
        in: query
        name: locale
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              items:
                items:
                  type: object
                type: array
            type: object
      summary: List published pages
      tags:
      - Pages
  /public/pages/{slug}:
    get:
      description: Returns a single published page with config and SEO data
      parameters:
      - description: Page slug
        in: path
        name: slug
        required: true
        type: string
      - description: Locale (zh or en)
        in: query
        name: locale
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get page by slug
      tags:
      - Pages
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}" for JWT authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"